Fixed dropout.

+++ libnd4j/include/ops/declarable/generic/parity_ops/dropout.cpp
@@ -38,7 +38,7 @@ CONFIGURABLE_OP_IMPL(dropout, 1, 1, true, 1, 1) {
     int seed = INT_ARG(0);
 
     // FIXME: float?
-    float probValue = T_ARG(0);
+    double probValue = T_ARG(0);
     if (block.width() > 1)
         reduceShape = INPUT_VARIABLE(1);
